Working around MDAnalysis's way of writing trajectory files
While you can go through a trajectory time step by time step
pulling out each frame you cannot use that to write a trajectory
directly (in another format for example). You have to select some
atoms to create an AtomGroup which gets update every time you
read a time step. You have to pass the AtomGroup to the trajectory
writer. If you pass a time step to the trajectory writer then
the code crashes complaining that the object you passed is neither
an AtomGroup nor a Universe. To make matters worse the MDAnalysis
documentation is full of broken examples that pass time steps to
the trajectory writer. This page
https://www.mdanalysis.org/MDAnalysisTutorial/writing.html
is the only place where I found the correct way of doing this.
The other interesting question is how "select_atoms" works.
Thankfully the selection "all" seems to work.